Final Project has few bugs
The app only works on Pixel 2XL API30 for now.

I would like to try to make a updateButton() method that easily calculates the buttons to be updated for the other methods, but I can't get it to work no matter what.

a weird bug is when I press study timer, start timer, and then reset. This causes the timer to reset back to 10 minutes and the "breakStudyTimer" does not update.